requirements.txt
streamlit
langchain
boto3                # AWS SDK to connect to Bedrock
langchain-community
python-dotenv        # To manage your API keys safely
mermaid-python       # For visual flowcharts
langchain-text-splitters
pypdf